initActivityPanels: Resolve scoping issue for loop iterator
The let	statement declares a block scope local variable.
Previously, the loop iterator (i) was not visible within the bound click function.
Changing the declaration of i to var opened up the scope to allow the iterator
to be seen incorrectly within the bound click function. Updating to create a new
scope for the iterator.