Remove readyVideos()
When preload is set to none, the readyVideos() behavior seems to be
interfering with the normal video loading behavior in the phonegap
applications on the lenovo and asus. desktop-chrome seems unaffected and
works well either way. This change makes the videos behave well
everywhere, as far as my testing has seen.